+++ src/automatic_differentiation/derivative/scalar_valued.rs
@@ -156,9 +156,10 @@ macro_rules! get_sderivative {
         ///
         /// # Arguments
         ///
-        /// `x0` - Evaluation point, `x₀ ∈ ℝ`.
-        /// `p` - Parameter vector. This is a vector of additional runtime parameters that the
-        ///       function may depend on but is not differentiated with respect to.
+        /// * `x0` - Evaluation point, `x₀ ∈ ℝ`.
+        /// * `p` - Extra runtime parameter. This is a parameter (can be of any arbitrary type)
+        ///   defined at runtime that the function may depend on but is not differentiated with
+        ///   respect to.
         ///
         /// # Returns
         ///